phase-63: HS-63-01 the meeting models — meeting_session becomes a package, the data layer carved verbatim
The five pure dataclasses (Bookmark, TranscriptSegment, IntelSnapshot,
MeetingSaveResult, MeetingState) + their helpers move to
holdspeak/meeting_session/models.py with a body-line-diff verbatim
proof (0 original lines lost); session.py keeps the machinery untouched
except import depth; __init__ re-exports every public name so all 38
importing test files and 5 production sites work unchanged — ZERO test
edits. The brief planned holdspeak/meeting/ but MeetingRecorder owns
that module name; the package conversion keeps the canonical import
point instead.

Trap caught mid-story: indented try-block optional imports silently
swallowed the packaging mistake (from .intel resolved against the
package, ImportError, intel quietly None) — 8 tests failed loudly and
the all-indentation fix landed; lesson recorded for stories 02-04.

Full suite 2768 passed, 17 skipped.
