+++ pm/roadmap/holdspeak/phase-63-backend-decomposition/evidence-story-01.md
@@ -1,0 +1,50 @@
+# Evidence — HS-63-01: The meeting models
+
+**Date:** 2026-06-12
+**Verdict:** done. `meeting_session.py` is a package; the five pure
+dataclasses live in `models.py`; every import in the repo and all 38 test
+files work unchanged. **Zero test edits.**
+
+## What shipped
+
+- `holdspeak/meeting_session/` (the module → package conversion; the
+  canonical import point is unchanged):
+  - `models.py` (240 lines): `Bookmark`, `TranscriptSegment`,
+    `IntelSnapshot`, `MeetingSaveResult`, `MeetingState`, plus the
+    `_device_descriptor_to_dict` / `_iso_or_none` helpers — bodies
+    verbatim.
+  - `session.py` (1,460): everything else, untouched except the import
+    block (sibling-relative imports promoted one level, `.models`
+    added).
+  - `__init__.py` (26): the re-exports (including
+    `_device_descriptor_to_dict`, which `web_runtime` and
+    `routes/system.py` import directly).
+- The collision found at execution: `holdspeak/meeting.py`
+  (MeetingRecorder) already owns the `meeting` name, so the brief's
+  planned `holdspeak/meeting/` package became the
+  `meeting_session/`-as-package conversion instead — strictly better,
+  since the import point never moves.
+
+## The verbatim proof
+
+A body-line diff (imports/blanks/docstrings excluded) between the
+original file and (models.py + session.py): **0 original lines lost**;
+the only additions are the models docstring and the `.models` import
+block.
+
+## A trap found and fixed mid-story
+
+The first relative-import fix used a column-0 regex and missed the
+INDENTED optional imports inside `try:` blocks (`from .intel import …`),
+which then resolved against the package, raised ImportError, and were
+silently swallowed by the existing `except ImportError` fallback — intel
+quietly became `None` and 8 tests failed loudly (the suite did its job).
+All-indentation fix applied; the lesson (try/except import fallbacks
+mask packaging mistakes) recorded for stories 02–04.
+
+## Proof
+
+- Full suite: **2768 passed, 17 skipped** — identical to the Phase-62
+  close, with zero test files touched.
+- Import smoke: every consumer (`web_runtime`, `meeting_import`,
+  `intel_queue`, the routes) imports clean.
